PCOCamera is now lazily reprogrammed for manual mode acquisition.
Previously the camera was always programmed for manual mode acquisition after a shot, then reprogrammed for buffered acquisition before the next shot. Doing that slows things down and is only necessary if "Snap" and/or "Continuous" in the blacs tab is used bewteen the shots. Now the camera is only configured for manual acquisition once "Snap" or "Continuos" is used, otherwise it just stays configured for buffered acquisition to save time.